// src/web/lib/api.ts
export class LootBoxDto implements ILootBoxDto {
    created_at!: Date;
    id!: number;
    is_open!: boolean;
    player!: PlayerPartialDto;
    result?: CreatorPartialDto | null;
    updated_at!: Date;

    constructor(data?: ILootBoxDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.player = new PlayerPartialDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.created_at = _data["created_at"] ? new Date(_data["created_at"].toString()) : null as any;
            this.id = _data["id"] !== undefined ? _data["id"] : null as any;
            this.is_open = _data["is_open"] !== undefined ? _data["is_open"] : null as any;
            this.player = _data["player"] ? PlayerPartialDto.fromJS(_data["player"]) : new PlayerPartialDto();
            this.result = _data["result"] ? CreatorPartialDto.fromJS(_data["result"]) : null as any;
            this.updated_at = _data["updated_at"] ? new Date(_data["updated_at"].toString()) : null as any;
        }
    }

    static fromJS(data: any): LootBoxDto {
        data = typeof data === 'object' ? data : {};
        let result = new LootBoxDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["created_at"] = this.created_at ? this.created_at.toISOString() : null as any;
        data["id"] = this.id !== undefined ? this.id : null as any;
        data["is_open"] = this.is_open !== undefined ? this.is_open : null as any;
        data["player"] = this.player ? this.player.toJSON() : null as any;
        data["result"] = this.result ? this.result.toJSON() : null as any;
        data["updated_at"] = this.updated_at ? this.updated_at.toISOString() : null as any;
        return data;
    }
}
// ...
export class PlayerPartialDto implements IPlayerPartialDto {
    avatar_url!: string;
    created_at!: Date;
    credits!: number;
    id!: number;
    name!: string;
    platform!: Platform;
    platform_id!: string;
    platform_url!: string;
    portfolio!: number;
    slug!: string;
    type!: PlayerType;
    updated_at!: Date;
    value!: number;

    constructor(data?: IPlayerPartialDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.avatar_url = _data["avatar_url"] !== undefined ? _data["avatar_url"] : null as any;
            this.created_at = _data["created_at"] ? new Date(_data["created_at"].toString()) : null as any;
            this.credits = _data["credits"] !== undefined ? _data["credits"] : null as any;
            this.id = _data["id"] !== undefined ? _data["id"] : null as any;
            this.name = _data["name"] !== undefined ? _data["name"] : null as any;
            this.platform = _data["platform"] !== undefined ? _data["platform"] : null as any;
            this.platform_id = _data["platform_id"] !== undefined ? _data["platform_id"] : null as any;
            this.platform_url = _data["platform_url"] !== undefined ? _data["platform_url"] : null as any;
            this.portfolio = _data["portfolio"] !== undefined ? _data["portfolio"] : null as any;
            this.slug = _data["slug"] !== undefined ? _data["slug"] : null as any;
            this.type = _data["type"] !== undefined ? _data["type"] : null as any;
            this.updated_at = _data["updated_at"] ? new Date(_data["updated_at"].toString()) : null as any;
            this.value = _data["value"] !== undefined ? _data["value"] : null as any;
        }
    }

    static fromJS(data: any): PlayerPartialDto {
        data = typeof data === 'object' ? data : {};
        let result = new PlayerPartialDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["avatar_url"] = this.avatar_url !== undefined ? this.avatar_url : null as any;
        data["created_at"] = this.created_at ? this.created_at.toISOString() : null as any;
        data["credits"] = this.credits !== undefined ? this.credits : null as any;
        data["id"] = this.id !== undefined ? this.id : null as any;
        data["name"] = this.name !== undefined ? this.name : null as any;
        data["platform"] = this.platform !== undefined ? this.platform : null as any;
        data["platform_id"] = this.platform_id !== undefined ? this.platform_id : null as any;
        data["platform_url"] = this.platform_url !== undefined ? this.platform_url : null as any;
        data["portfolio"] = this.portfolio !== undefined ? this.portfolio : null as any;
        data["slug"] = this.slug !== undefined ? this.slug : null as any;
        data["type"] = this.type !== undefined ? this.type : null as any;
        data["updated_at"] = this.updated_at ? this.updated_at.toISOString() : null as any;
        data["value"] = this.value !== undefined ? this.value : null as any;
        return data;
    }
}
// ...
export class PlayerDto extends PlayerPartialDto implements IPlayerDto {
    history!: PortfolioSnapshotDto[];
    loot_boxes!: LootBoxDto[];
    shares!: PlayerShareDto[];
    transactions!: PlayerTransactionDto[];

    constructor(data?: IPlayerDto) {
        super(data);
        if (!data) {
            this.history = [];
            this.loot_boxes = [];
            this.shares = [];
            this.transactions = [];
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            if (Array.isArray(_data["history"])) {
                this.history = [] as any;
                for (let item of _data["history"])
                    this.history!.push(PortfolioSnapshotDto.fromJS(item));
            }
            else {
                this.history = null as any;
            }
            if (Array.isArray(_data["loot_boxes"])) {
                this.loot_boxes = [] as any;
                for (let item of _data["loot_boxes"])
                    this.loot_boxes!.push(LootBoxDto.fromJS(item));
            }
            else {
                this.loot_boxes = null as any;
            }
            if (Array.isArray(_data["shares"])) {
                this.shares = [] as any;
                for (let item of _data["shares"])
                    this.shares!.push(PlayerShareDto.fromJS(item));
            }
            else {
                this.shares = null as any;
            }
            if (Array.isArray(_data["transactions"])) {
                this.transactions = [] as any;
                for (let item of _data["transactions"])
                    this.transactions!.push(PlayerTransactionDto.fromJS(item));
            }
            else {
                this.transactions = null as any;
            }
        }
    }

    static override fromJS(data: any): PlayerDto {
        data = typeof data === 'object' ? data : {};
        let result = new PlayerDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.history)) {
            data["history"] = [];
            for (let item of this.history)
                data["history"].push(item ? item.toJSON() : null as any);
        }
        if (Array.isArray(this.loot_boxes)) {
            data["loot_boxes"] = [];
            for (let item of this.loot_boxes)
                data["loot_boxes"].push(item ? item.toJSON() : null as any);
        }
        if (Array.isArray(this.shares)) {
            data["shares"] = [];
            for (let item of this.shares)
                data["shares"].push(item ? item.toJSON() : null as any);
        }
        if (Array.isArray(this.transactions)) {
            data["transactions"] = [];
            for (let item of this.transactions)
                data["transactions"].push(item ? item.toJSON() : null as any);
        }
        super.toJSON(data);
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Players.PlayerDto`. */
export interface IPlayerDto extends IPlayerPartialDto {
    history: PortfolioSnapshotDto[];
    loot_boxes: LootBoxDto[];
    shares: PlayerShareDto[];
    transactions: PlayerTransactionDto[];
}

/** Mirrors `Services.Players.PlayerOrderBy`. */
export enum PlayerOrderBy {
    Name = "Name",
    Credits = "Credits",
    Portfolio = "Portfolio",
}

/** Mirrors `TTX.App.Dto.Players.PlayerShareDto`. */
export class PlayerShareDto implements IPlayerShareDto {
    creator!: CreatorPartialDto;
    quantity!: number;

    constructor(data?: IPlayerShareDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.creator = new CreatorPartialDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.creator = _data["creator"] ? CreatorPartialDto.fromJS(_data["creator"]) : new CreatorPartialDto();
            this.quantity = _data["quantity"] !== undefined ? _data["quantity"] : null as any;
        }
    }

    static fromJS(data: any): PlayerShareDto {
        data = typeof data === 'object' ? data : {};
        let result = new PlayerShareDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["creator"] = this.creator ? this.creator.toJSON() : null as any;
        data["quantity"] = this.quantity !== undefined ? this.quantity : null as any;
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Players.PlayerShareDto`. */
export interface IPlayerShareDto {
    creator: CreatorPartialDto;
    quantity: number;
}

/** Mirrors `TTX.App.Dto.Transactions.PlayerTransactionDto`. */
export class PlayerTransactionDto extends TransactionDto implements IPlayerTransactionDto {
    creator!: CreatorPartialDto;

    constructor(data?: IPlayerTransactionDto) {
        super(data);
        if (!data) {
            this.creator = new CreatorPartialDto();
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.creator = _data["creator"] ? CreatorPartialDto.fromJS(_data["creator"]) : new CreatorPartialDto();
        }
    }

    static override fromJS(data: any): PlayerTransactionDto {
        data = typeof data === 'object' ? data : {};
        let result = new PlayerTransactionDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["creator"] = this.creator ? this.creator.toJSON() : null as any;
        super.toJSON(data);
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Transactions.PlayerTransactionDto`. */
export interface IPlayerTransactionDto extends ITransactionDto {
    creator: CreatorPartialDto;
}

/** Mirrors `TTX.Domain.Models.PlayerType`. */
export enum PlayerType {
    User = "User",
    Admin = "Admin",
}

/** Mirrors `TTX.App.Dto.Players.PortfolioSnapshotDto`. */
export class PortfolioSnapshotDto implements IPortfolioSnapshotDto {
    player_id!: number;
    time!: Date;
    value!: number;

    constructor(data?: IPortfolioSnapshotDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.player_id = _data["player_id"] !== undefined ? _data["player_id"] : null as any;
            this.time = _data["time"] ? new Date(_data["time"].toString()) : null as any;
            this.value = _data["value"] !== undefined ? _data["value"] : null as any;
        }
    }

    static fromJS(data: any): PortfolioSnapshotDto {
        data = typeof data === 'object' ? data : {};
        let result = new PortfolioSnapshotDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["player_id"] = this.player_id !== undefined ? this.player_id : null as any;
        data["time"] = this.time ? this.time.toISOString() : null as any;
        data["value"] = this.value !== undefined ? this.value : null as any;
        return data;
    }
}